pyproject.toml
[tool.poetry]
name = "decrypt-otpauth-files"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
bpylist = "^0.1.4"
click = "^8.0.3"
Pillow = "^9.0.1"
qrcode = "^7.3.1"
Jinja2 = "^3.0.3"
weasyprint = "^54.1"
GitPython = "^3.1.26"
rncryptor = "^3.3.0"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
